# backend/srcs/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST du projet Pan Bagnat.",
        "title": "Pan Bagnat API",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/roles": {
            "get": {
                "description": "Returns all available roles for your campus, with optional filtering, sorting, and pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Get Role List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter expression (e.g. \\",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination token for the next page",
                        "name": "next_page_token",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order: asc or desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roles.RoleGetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Imports a new role with the given name, color, default status, and module assignments.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Create Role",
                "parameters": [
                    {
                        "description": "Role creation payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roles.RoleCreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly created role",
                        "schema": {
                            "$ref": "#/definitions/api.Role"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON input or missing required fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/roles/{roleID}": {
            "get": {
                "description": "Returns details about a specific role and its associated modules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Get Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Role"
                        }
                    },
                    "400": {
                        "description": "Role ID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the specified role for your campus (all role data will be lost).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Delete Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid role ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the name, color, and/or default status of an existing role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Patch Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roles.RolePatchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated role",
                        "schema": {
                            "$ref": "#/definitions/api.Role"
                        }
                    },
                    "400": {
                        "description": "Invalid role ID or JSON body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Module": {
            "type": "object",
            "properties": {
                "git_branch": {
                    "description": "GitBranch is the branch currently tracked for deployments",
                    "type": "string",
                    "example": "main"
                },
                "git_url": {
                    "description": "GitURL is the repository URL where the module’s source lives",
                    "type": "string",
                    "example": "https://github.com/some-user/some-repo"
                },
                "icon_url": {
                    "description": "IconURL is the link to the module’s icon or logo",
                    "type": "string",
                    "example": "https://someURL/image.png"
                },
                "id": {
                    "description": "ID is the unique identifier of the module",
                    "type": "string",
                    "example": "module_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "last_update": {
                    "description": "LastUpdate is the UTC timestamp of the module’s last update check",
                    "type": "string",
                    "example": "2025-02-18T15:00:00Z"
                },
                "late_commits": {
                    "description": "LateCommits is the number of commits behind the latest version",
                    "type": "integer",
                    "example": 2
                },
                "latest_version": {
                    "description": "LatestVersion is the most recent version available upstream",
                    "type": "string",
                    "example": "1.7"
                },
                "name": {
                    "description": "Name is the human-readable title of the module",
                    "type": "string",
                    "example": "Captain Hook"
                },
                "roles": {
                    "description": "Roles lists the roles that have access to this module",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Role"
                    }
                },
                "slug": {
                    "description": "Slug is the URL-friendly identifier for the module",
                    "type": "string",
                    "example": "captain-hook-main"
                },
                "ssh_public_key": {
                    "description": "SSHPublicKey is the RSA public key used for module deployments",
                    "type": "string",
                    "example": "ssh-rsa AAAA..."
                },
                "status": {
                    "description": "Status indicates whether the module is enabled or disabled",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ModuleStatus"
                        }
                    ],
                    "example": "enabled"
                },
                "version": {
                    "description": "Version is the currently deployed version of the module",
                    "type": "string",
                    "example": "1.2"
                }
            }
        },
...
        "api.ModuleStatus": {
            "type": "string",
            "enum": [
                "cloning",
                "waiting_for_deploy",
                "disabled",
                "enabled"
            ],
            "x-enum-varnames": [
                "Cloning",
                "WaitingDeploy",
                "Disabled",
                "Enabled"
            ]
        },
        "api.Role": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "Color is the hex color code associated with the role in the UI",
                    "type": "string",
                    "example": "#FF00FF"
                },
                "id": {
                    "description": "ID is the unique identifier of the role",
                    "type": "string",
                    "example": "role_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "is_default": {
                    "description": "IsDefault indicates whether this role is assigned by default to new users",
                    "type": "boolean",
                    "example": true
                },
                "modules": {
                    "description": "Modules lists the Module objects this role grants access to",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Module"
                    }
                },
                "name": {
                    "description": "Name is the human‐readable name of the role (e.g. \"IT\", \"Admin\")",
                    "type": "string",
                    "example": "IT"
                },
                "users": {
                    "description": "Users lists the User objects that have this role",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.User"
                    }
                },
                "usersCount": {
                    "description": "UsersCount is the total number of users assigned to this role",
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "api.User": {
            "type": "object",
            "properties": {
                "ft_id": {
                    "description": "FtID is the numeric 42-intranet ID of the user",
                    "type": "integer",
                    "example": 1492
                },
                "ft_is_staff": {
                    "description": "FtIsStaff indicates whether the user is a 42-intranet staff member",
                    "type": "boolean",
                    "example": true
                },
                "ft_login": {
                    "description": "FtLogin is the 42-intranet login handle (e.g. \"heinz\")",
                    "type": "string",
                    "example": "heinz"
                },
                "ft_photo": {
                    "description": "PhotoURL is the URL to the user’s 42-intranet profile picture",
                    "type": "string",
                    "example": "https://intra.42.fr/some-login/some-id"
                },
                "id": {
                    "description": "ID is the unique identifier of the user",
                    "type": "string",
                    "example": "user_01HZ0MMK4S6VQW4WPHB6NZ7R7X"
                },
                "is_staff": {
                    "description": "IsStaff indicates whether the user has staff privileges within Pan Bagnat",
                    "type": "boolean",
                    "example": true
                },
                "last_seen": {
                    "description": "LastSeen is the UTC timestamp of the user’s last activity",
                    "type": "string",
                    "example": "2025-02-18T15:00:00Z"
                },
                "roles": {
                    "description": "Roles lists the roles assigned to the user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Role"
                    }
                }
            }
        },
...
        "roles.RoleCreateInput": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "Color is the hex color code for the role.",
                    "type": "string",
                    "example": "#00FF00"
                },
                "is_default": {
                    "description": "IsDefault indicates whether this role is the default for new users.",
                    "type": "boolean",
                    "example": false
                },
                "modules": {
                    "description": "Modules lists the IDs of modules this role should have access to.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"module_01\"",
                        "\"module_02\"]"
                    ]
                },
                "name": {
                    "description": "Name is the human-readable name of the role.",
                    "type": "string",
                    "example": "Support"
                }
            }
        },
        "roles.RoleGetResponse": {
            "type": "object",
            "properties": {
                "next_page_token": {
                    "description": "NextPageToken is the token to retrieve the next page of results.",
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "roles": {
                    "description": "Roles is the list of roles on this page.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Role"
                    }
                }
            }
        },
        "roles.RolePatchInput": {
            "description": "API Role model",
            "type": "object",
            "properties": {
                "color": {
                    "description": "Color is the new hex color code for the role.",
                    "type": "string",
                    "example": "#00FF00"
                },
                "is_default": {
                    "description": "IsDefault toggles whether this role is the default for new users.",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "Name is the new name for the role.",
                    "type": "string",
                    "example": "Support"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    },
...
}